# doc/swagger.yaml
definitions:
...
  crypto.SignatureAlgorithm:
    enum:
    - EdDSA
    - ES256K
    - ES256
    - ES384
    - PS256
    - Dilithium2
    - Dilithium3
    - Dilithium5
    type: string
    x-enum-varnames:
    - EdDSA
    - ES256K
    - ES256
    - ES384
    - PS256
    - Dilithium2Sig
    - Dilithium3Sig
    - Dilithium5Sig
...
  exchange.ClaimFormat:
    properties:
      jwt:
        $ref: '#/definitions/exchange.JWTType'
      jwt_vc:
        $ref: '#/definitions/exchange.JWTType'
      jwt_vp:
        $ref: '#/definitions/exchange.JWTType'
      ldp:
        $ref: '#/definitions/exchange.LDPType'
      ldp_vc:
        $ref: '#/definitions/exchange.LDPType'
      ldp_vp:
        $ref: '#/definitions/exchange.LDPType'
    type: object
...
  exchange.InputDescriptor:
    properties:
      constraints:
        $ref: '#/definitions/exchange.Constraints'
      format:
        $ref: '#/definitions/exchange.ClaimFormat'
      group:
        description: Must match a grouping strings listed in the `from` values of
          a submission requirement rule
        items:
          type: string
        type: array
      id:
        description: Must be unique within the Presentation Definition
        type: string
      name:
        type: string
      purpose:
        description: Purpose for which claim's data is being requested
        type: string
    required:
    - constraints
    - id
    type: object
  exchange.JWTType:
    properties:
      alg:
        items:
          $ref: '#/definitions/crypto.SignatureAlgorithm'
        type: array
    required:
    - alg
    type: object
  exchange.LDPType:
    properties:
      proof_type:
        items:
          type: string
        type: array
    required:
    - proof_type
    type: object
...
  exchange.PresentationDefinition:
    properties:
      format:
        $ref: '#/definitions/exchange.ClaimFormat'
      frame:
        description: https://identity.foundation/presentation-exchange/#json-ld-framing-feature
      id:
        type: string
      input_descriptors:
        items:
          $ref: '#/definitions/exchange.InputDescriptor'
        type: array
      name:
        type: string
      purpose:
        type: string
      submission_requirements:
        items:
          $ref: '#/definitions/exchange.SubmissionRequirement'
        type: array
    required:
    - id
    - input_descriptors
    type: object
...
  exchange.Selection:
    enum:
    - all
    - pick
    type: string
    x-enum-varnames:
    - All
    - Pick
...
  exchange.SubmissionRequirement:
    properties:
      count:
        minimum: 1
        type: integer
      from:
        type: string
      from_nested:
        items:
          $ref: '#/definitions/exchange.SubmissionRequirement'
        type: array
      max:
        type: integer
      min:
        type: integer
      name:
        type: string
      purpose:
        type: string
      rule:
        $ref: '#/definitions/exchange.Selection'
    required:
    - rule
    type: object
...
  github_com_cyware_ssi-service_pkg_service_issuance.ClaimTemplates:
    additionalProperties: {}
    type: object
  github_com_cyware_ssi-service_pkg_service_issuance.CredentialTemplate:
    properties:
      credentialInputDescriptor:
        description: |-
          Optional.
          When present, it's the ID of the input descriptor in the application. Corresponds to one of the
          PresentationDefinition.InputDescriptors[].ID in the credential manifest. When creating a credential, the base
          data will be populated from the provided submission that matches this ID.
          When absent, there will be no base data for the credentials created. Additionally, no JSON path strings in
          ClaimTemplates.Data will be resolved.
        type: string
      data:
        allOf:
        - $ref: '#/definitions/github_com_cyware_ssi-service_pkg_service_issuance.ClaimTemplates'
        description: |-
          Data that will be used to determine credential claims.
          Values may be json path like strings, or any other JSON primitive. Each entry will be used to come up with a
          claim about the credentialSubject in the credential that will be issued.
      expiry:
        allOf:
        - $ref: '#/definitions/github_com_cyware_ssi-service_pkg_service_issuance.TimeLike'
        description: Parameter to determine the expiry of the credential.
      id:
        description: ID corresponding to an OutputDescriptor.ID from the manifest.
        type: string
      revocable:
        description: Whether the credentials created should be revocable.
        type: boolean
      schema:
        description: ID of the CredentialSchema to be used for the issued credential.
        type: string
    type: object
  github_com_cyware_ssi-service_pkg_service_issuance.Template:
    properties:
      credentialManifest:
        description: ID of the credential manifest that this template corresponds
          to.
        type: string
      credentials:
        description: Info required to create a credential from a credential application.
        items:
          $ref: '#/definitions/github_com_cyware_ssi-service_pkg_service_issuance.CredentialTemplate'
        type: array
      id:
        description: ID of this template.
        type: string
      issuer:
        description: ID of the issuer that will be issuing the credentials.
        type: string
      verificationMethodId:
        description: |-
          The id of the verificationMethod (see https://www.w3.org/TR/did-core/#verification-methods) who's privateKey is
          stored in ssi-service. The verificationMethod must be part of the did document associated with `issuer`.
          The private key associated with the verificationMethod's publicKey will be used to sign the credentials.
        example: did:key:z6MkkZDjunoN4gyPMx5TSy7Mfzw22D2RZQZUcx46bii53Ex3#z6MkkZDjunoN4gyPMx5TSy7Mfzw22D2RZQZUcx46bii53Ex3
        type: string
    required:
    - credentialManifest
    - issuer
    - verificationMethodId
    type: object
  github_com_cyware_ssi-service_pkg_service_issuance.TimeLike:
    properties:
      duration:
        allOf:
        - $ref: '#/definitions/time.Duration'
        description: For a fixed offset from when it was issued.
      time:
        description: For fixed time in the future.
        type: string
    type: object
...
  github_com_cyware_ssi-service_pkg_service_well-known.VerifyDIDConfigurationRequest:
    properties:
      origin:
        description: |-
          Represents an origin to fetch the DID Configuration Resource from. Must be serialized as described in https://html.spec.whatwg.org/multipage/browsers.html#origin.
          The `scheme` MUST be `https`.
        example: https://example.com
        type: string
    required:
    - origin
    type: object
  github_com_cyware_ssi-service_pkg_service_well-known.VerifyDIDConfigurationResponse:
    properties:
      didConfiguration:
        description: The configuration that was fetched from the origin's well known
          location.
        type: string
      reason:
        description: When Verified == false, the reason why it wasn't verified.
        type: string
      verified:
        description: Whether the DIDConfiguration was verified.
        type: boolean
    type: object
...
  manifest.CredentialManifest:
    properties:
      description:
        type: string
      format:
        $ref: '#/definitions/exchange.ClaimFormat'
      id:
        type: string
      issuer:
        $ref: '#/definitions/manifest.Issuer'
      name:
        type: string
      output_descriptors:
        items:
          $ref: '#/definitions/manifest.OutputDescriptor'
        type: array
      presentation_definition:
        $ref: '#/definitions/exchange.PresentationDefinition'
      spec_version:
        type: string
    required:
    - id
    - issuer
    - output_descriptors
    - spec_version
    type: object
...
  manifest.Issuer:
    properties:
      id:
        type: string
      name:
        type: string
      styles:
        allOf:
        - $ref: '#/definitions/rendering.EntityStyleDescriptor'
        description: |-
          an object or URI as defined by the DIF Entity Styles specification
          https://identity.foundation/wallet-rendering/#entity-styles
    required:
    - id
    type: object
  manifest.OutputDescriptor:
    properties:
      description:
        type: string
      display:
        allOf:
        - $ref: '#/definitions/rendering.DataDisplay'
        description: 'both below: an object or URI as defined by the DIF Entity Styles
          specification'
      id:
        description: Must be unique within a manifest
        type: string
      name:
        type: string
      schema:
        type: string
      styles:
        $ref: '#/definitions/rendering.EntityStyleDescriptor'
    required:
    - id
    - schema
    type: object
...
  pkg_server_router.CreateDIDConfigurationRequest:
    properties:
      expirationDate:
        description: Will be used to set the `DomainLinkageCredential.credentialSubject.expirationDate`.
          Required.
        example: "2051-10-05T14:48:00.000Z"
        type: string
      issuanceDate:
        description: |-
          Will be used to set the `DomainLinkageCredential.credentialSubject.issuanceDate`. If left empty, then the current
          time will be used.
          Optional.
        example: "2021-10-05T14:48:00.000Z"
        type: string
      issuerDid:
        description: |-
          DID that identifies who the issuer of the credential(s) will be.
          Required.
        example: did:key:z6MkkZDjunoN4gyPMx5TSy7Mfzw22D2RZQZUcx46bii53Ex3
        type: string
      origin:
        description: |-
          Serialization of an origin as described in https://html.spec.whatwg.org/multipage/browsers.html#origin. Represents
          the origin that the IssuerDID controls, which will be included in the `DomainLinkageCredential.credentialSubject.origin`
          value.
          Required.
        example: https://www.cyware.github.io/
        type: string
      verificationMethodId:
        description: |-
          The id of the verificationMethod (see https://www.w3.org/TR/did-core/#verification-methods) who's privateKey is
          stored in ssi-service. The verificationMethod must be part of the did document associated with `issuer`.
          The private key associated with the verificationMethod's publicKey will be used to sign the domain linkage credential.
          Required.
        example: did:key:z6MkkZDjunoN4gyPMx5TSy7Mfzw22D2RZQZUcx46bii53Ex3#z6MkkZDjunoN4gyPMx5TSy7Mfzw22D2RZQZUcx46bii53Ex3
        type: string
    required:
    - expirationDate
    - issuerDid
    - origin
    - verificationMethodId
    type: object
  pkg_server_router.CreateDIDConfigurationResponse:
    properties:
      didConfiguration:
        allOf:
        - $ref: '#/definitions/pkg_server_router.DIDConfiguration'
        description: The DID Configuration Resource according to https://identity.foundation/.well-known/resources/did-configuration/#did-configuration-resource
      wellKnownLocation:
        description: The location in which the `didConfiguration` value should be
          hosted.
        type: string
    type: object
  pkg_server_router.CreateIssuanceTemplateRequest:
    properties:
      credentialManifest:
        description: ID of the credential manifest that this template corresponds
          to.
        type: string
      credentials:
        description: Info required to create a credential from a credential application.
        items:
          $ref: '#/definitions/github_com_cyware_ssi-service_pkg_service_issuance.CredentialTemplate'
        type: array
      id:
        description: ID of this template.
        type: string
      issuer:
        description: ID of the issuer that will be issuing the credentials.
        type: string
      verificationMethodId:
        description: |-
          The id of the verificationMethod (see https://www.w3.org/TR/did-core/#verification-methods) who's privateKey is
          stored in ssi-service. The verificationMethod must be part of the did document associated with `issuer`.
          The private key associated with the verificationMethod's publicKey will be used to sign the credentials.
        example: did:key:z6MkkZDjunoN4gyPMx5TSy7Mfzw22D2RZQZUcx46bii53Ex3#z6MkkZDjunoN4gyPMx5TSy7Mfzw22D2RZQZUcx46bii53Ex3
        type: string
    required:
    - credentialManifest
    - issuer
    - verificationMethodId
    type: object
  pkg_server_router.CreateManifestRequest:
    properties:
      description:
        description: |-
          Explains what the Manifest in question is generally offering in exchange for meeting its requirements.
          Optional.
        type: string
      format:
        allOf:
        - $ref: '#/definitions/exchange.ClaimFormat'
        description: |-
          Formats that the issuer can support when issuing the credential. At least one needs to be set. We currently only
          support `jwt_vc` for issuance. See https://identity.foundation/claim-format-registry/#registry for the definition.
          TODO: support different claim formats https://github.com/cyware/ssi-service/issues/96
      issuerDid:
        description: |-
          DID that identifies who the issuer of the credential(s) will be.
          Required.
        type: string
      issuerName:
        description: |-
          Human-readable name the Issuer wishes to be recognized by.
          Optional.
        type: string
      name:
        description: |-
          Summarizing title for the Manifest in question.
          Optional.
        type: string
      outputDescriptors:
        description: Array of objects as defined in https://identity.foundation/credential-manifest/#output-descriptor.
        items:
          $ref: '#/definitions/manifest.OutputDescriptor'
        type: array
      presentationDefinition:
        allOf:
        - $ref: '#/definitions/exchange.PresentationDefinition'
        description: value of the presentation definition to use. Must be empty if
          `id` is present.
      presentationDefinitionId:
        description: id of the presentation definition created with the PresentationDefinitions
          API. Must be empty if `value` is present.
        type: string
      verificationMethodId:
        description: |-
          The id of the verificationMethod (see https://www.w3.org/TR/did-core/#verification-methods) who's privateKey is
          stored in ssi-service. The verificationMethod must be part of the did document associated with `issuer`.
          The private key associated with the verificationMethod's publicKey will be used to sign the issued credentials.
          Required.
        example: did:key:z6MkkZDjunoN4gyPMx5TSy7Mfzw22D2RZQZUcx46bii53Ex3#z6MkkZDjunoN4gyPMx5TSy7Mfzw22D2RZQZUcx46bii53Ex3
        type: string
    required:
    - format
    - issuerDid
    - outputDescriptors
    - verificationMethodId
    type: object
...
  pkg_server_router.CreateManifestResponse:
    properties:
      credential_manifest:
        $ref: '#/definitions/manifest.CredentialManifest'
    type: object
...
  pkg_server_router.DIDConfiguration:
    properties:
      '@context': {}
      linked_dids:
        items: {}
        type: array
    required:
    - '@context'
    - linked_dids
    type: object
...
  pkg_server_router.ListManifestResponse:
    properties:
      credential_manifest:
        $ref: '#/definitions/manifest.CredentialManifest'
      id:
        type: string
    type: object
  pkg_server_router.ListManifestsResponse:
    properties:
      manifests:
        items:
          $ref: '#/definitions/pkg_server_router.ListManifestResponse'
        type: array
    type: object
...
  rendering.ColorResource:
    properties:
      color:
        description: 'a HEX string color value (e.g. #00000)'
        type: string
    type: object
  rendering.DataDisplay:
    properties:
      description:
        $ref: '#/definitions/rendering.DisplayMappingObject'
      properties:
        items:
          $ref: '#/definitions/rendering.LabeledDisplayMappingObject'
        type: array
      subtitle:
        $ref: '#/definitions/rendering.DisplayMappingObject'
      title:
        $ref: '#/definitions/rendering.DisplayMappingObject'
    type: object
  rendering.DisplayMappingObject:
    properties:
      fallback:
        type: string
      path:
        description: |-
          Ifa path is present it must be an array of JSON Path string expressions
          and the schema property must also be present.
        items:
          type: string
        type: array
      schema:
        $ref: '#/definitions/rendering.DisplayMappingSchema'
      text:
        description: If path is not present, the text value is required with no other
          properties
        type: string
    type: object
...
  rendering.EntityStyleDescriptor:
    properties:
      background:
        $ref: '#/definitions/rendering.ColorResource'
      hero:
        $ref: '#/definitions/rendering.ImageResource'
      text:
        $ref: '#/definitions/rendering.ColorResource'
      thumbnail:
        $ref: '#/definitions/rendering.ImageResource'
    type: object
...
  rendering.LabeledDisplayMappingObject:
    properties:
      fallback:
        type: string
      label:
        type: string
      path:
        description: |-
          Ifa path is present it must be an array of JSON Path string expressions
          and the schema property must also be present.
        items:
          type: string
        type: array
      schema:
        $ref: '#/definitions/rendering.DisplayMappingSchema'
      text:
        description: If path is not present, the text value is required with no other
          properties
        type: string
    required:
    - label
    type: object
...
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
info:
  contact:
    email: tbd-developer@squareup.com
    name: CYWARE
    url: https://github.com/cyware/ssi-service/issues
  description: 'The Self Sovereign Identity Service: Managing DIDs, Verifiable Credentials,
    and more!'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: SSI Service API
  version: 0.0.3
paths:
...
  /v1/did-configurations:
    put:
      consumes:
      - application/json
      description: |-
        Creates a DID Configuration Resource which conforms to https://identity.foundation/.well-known/resources/did-configuration/#did-configuration-resource
        The `didConfiguration` can be hosted at the `wellKnownLocation` specified in the response.
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pkg_server_router.CreateDIDConfigurationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/pkg_server_router.CreateDIDConfigurationResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create DID Configurations
      tags:
      - DIDConfigurations
  /v1/did-configurations/verification:
    put:
      consumes:
      - application/json
      description: Verifies a DID Configuration Resource according to https://identity.foundation/.well-known/resources/did-configuration/#did-configuration-resource-verification
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_cyware_ssi-service_pkg_service_well-known.VerifyDIDConfigurationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_cyware_ssi-service_pkg_service_well-known.VerifyDIDConfigurationResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Verifies a DID Configuration Resource
      tags:
      - DIDConfigurations
...
  /v1/issuancetemplates:
    put:
      consumes:
      - application/json
      description: Creates an issuance template
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pkg_server_router.CreateIssuanceTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_cyware_ssi-service_pkg_service_issuance.Template'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create an issuance template
      tags:
      - IssuanceTemplates
  /v1/issuancetemplates/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an issuance template by its ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete an issuance template
      tags:
      - IssuanceTemplates
    get:
      consumes:
      - application/json
      description: Get an issuance template by its ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cyware_ssi-service_pkg_service_issuance.Template'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Get an issuance template
      tags:
      - IssuanceTemplates
...
  /v1/manifests:
    get:
      consumes:
      - application/json
      description: Checks for the presence of a query parameter and calls the associated
        filtered get method for Credential Manifests
      parameters:
      - description: string issuer
        in: query
        name: issuer
        type: string
      - description: string schema
        in: query
        name: schema
        type: string
      - description: string subject
        in: query
        name: subject
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.ListManifestsResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List Credential Manifests
      tags:
      - Manifests
    put:
      consumes:
      - application/json
      description: Create a Credential Manifest. Most fields map to the definitions
        from https://identity.foundation/credential-manifest/#general-composition.
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pkg_server_router.CreateManifestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/pkg_server_router.CreateManifestResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a Credential Manifest
      tags:
      - Manifests
...
swagger: "2.0"